+++ asmar-backend/src/main/java/com/aston/ecommerce/asmar/dao/ShippingInfoRepository.java
@@ -4,5 +4,5 @@ import com.aston.ecommerce.asmar.entity.ShippingInfo;
 import com.aston.ecommerce.asmar.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShippingInfoRepository extends JpaRepository<ShippingInfo, Integer> {
+public interface ShippingInfoRepository extends JpaRepository<ShippingInfo, Long> {
 }
